Refactored unit tests to address duplicate code and reduce imports
